refactor(telegram): split bot handler runtimes
